Silly int/double error in MPI reduce call fixed

+++ C++/src/JobAllocator.cpp
@@ -246,7 +246,8 @@ void LoadData(int id)
 		std::cout << "\tProcess " << ProcessRank << " has loaded in " << Data.size() << " datapoints in " << duration << std::endl; 
 	);
 	
-	double n = Data.size();
+	int n = Data.size();
+	
 	MPI_Reduce(&n,&TotalStars,1,MPI_INT,MPI_SUM,RootID,MPI_COMM_WORLD);
 	MPI_Reduce(&n, &MaxStarsInCore, 1,MPI_INT, MPI_MAX, RootID,MPI_COMM_WORLD);
